+++ threedi_model_migration/cli.py
@@ -1,21 +1,12 @@
 """Console script for threedi_model_migration."""
-from .application import clear_largefiles_cache
-from .application import download_inspect_plan
-from .conversion import repository_to_schematisations
-from .json_utils import custom_json_object_hook
-from .json_utils import custom_json_serializer
+from . import application
 from .metadata import load_metadata
 from .repository import DEFAULT_REMOTE
-from .repository import Repository
 
 import click
-import csv
-import dataclasses
-import json
 import logging
 import pathlib
 import re
-import shutil
 import sys
 
 
@@ -27,20 +18,14 @@ logger = logging.getLogger(__name__)
     "-b",
     "--base_path",
     type=click.Path(exists=True, readable=True, path_type=pathlib.Path),
+    default=pathlib.Path.cwd,
     help="Local root that contains the repositories",
-)
-@click.option(
-    "-n",
-    "--name",
-    type=str,
-    required=True,
-    help="The name of the repository (directory within path)",
 )
 @click.option(
     "-m",
     "--metadata_path",
     type=click.Path(exists=True, readable=True, path_type=pathlib.Path),
-    help="An optional path to a database dump of the modeldatabank",
+    help="An optional path to a database dump of the modeldatabank, required when using --uuid",
 )
 @click.option(
     "-v",
@@ -50,14 +35,10 @@ logger = logging.getLogger(__name__)
     help="Logging verbosity (0: error, 1: warning, 2: info, 3: debug)",
 )
 @click.pass_context
-def main(ctx, base_path, name, metadata_path, verbosity):
+def main(ctx, base_path, metadata_path, verbosity):
     """Console script for threedi_model_migration."""
-    if not base_path:
-        base_path = pathlib.Path.cwd()
-
     ctx.ensure_object(dict)
     ctx.obj["base_path"] = base_path
-    ctx.obj["repository"] = Repository(base_path, name)
     ctx.obj["inspection_path"] = base_path / "_inspection"
     ctx.obj["metadata_path"] = metadata_path
 
@@ -73,6 +54,10 @@ def main(ctx, base_path, name, metadata_path, verbosity):
 
 
 @main.command()
+@click.argument(
+    "slug",
+    type=str,
+)
 @click.option(
     "-r",
     "--remote",
@@ -87,39 +72,40 @@ def main(ctx, base_path, name, metadata_path, verbosity):
     default=False,
     help="Whether to use UUIDs instead of repository slugs in the remote",
 )
-@click.pass_context
-def download(ctx, remote, uuid):
-    """Clones / pulls a repository"""
-    repository = ctx.obj["repository"]
-    if uuid:
-        if not ctx.obj["metadata_path"]:
-            raise ValueError("Please supply metadata_path")
-        metadata = load_metadata(ctx.obj["metadata_path"])
-        remote_name = str(metadata[repository.slug].repo_uuid)
-    else:
-        remote_name = repository.slug
-
-    if remote.endswith("/"):
-        remote = remote[:-1]
-
-    repository.download(remote + "/" + remote_name)
-
-
-@main.command()
-@click.pass_context
-def delete(ctx):
-    """Clones / pulls a repository"""
-    repository = ctx.obj["repository"]
-    shutil.rmtree(repository.path)
-
-
-@main.command()
 @click.option(
-    "-i",
-    "--indent",
-    type=int,
-    default=4,
-    help="The indentation in case of JSON output",
+    "--lfclear/--no-lfclear",
+    type=bool,
+    default=False,
+    help="Whether to clear the largefiles usercache ($HOME/.cache/largefiles) afterwards",
+)
+@click.pass_context
+def download(ctx, slug, remote, uuid, lfclear):
+    """Clones / pulls a repository"""
+    application.download(
+        ctx.obj["base_path"],
+        slug,
+        remote,
+        uuid,
+        ctx.obj["metadata_path"],
+        lfclear,
+    )
+
+
+@main.command()
+@click.argument(
+    "slug",
+    type=str,
+)
+@click.pass_context
+def delete(ctx, slug):
+    """Removes a repository"""
+    application.delete(ctx.obj["base_path"], slug)
+
+
+@main.command()
+@click.argument(
+    "slug",
+    type=str,
 )
 @click.option(
     "-l",
@@ -134,53 +120,21 @@ def delete(ctx):
     default=False,
 )
 @click.pass_context
-def inspect(ctx, indent, last_update, quiet):
+def inspect(ctx, slug, last_update, quiet):
     """Inspects revisions, sqlites, and global settings in a repository"""
-    repository = ctx.obj["repository"]
-    inspection_path = ctx.obj["inspection_path"]
-
-    INSPECT_CSV_FIELDNAMES = [
-        "revision_nr",
-        "revision_hash",
-        "last_update",
-        "sqlite_path",
-        "settings_id",
-        "settings_name",
-    ]
-
-    if not quiet:
-        stdout = click.get_text_stream("stdout")
-        writer = csv.DictWriter(stdout, fieldnames=INSPECT_CSV_FIELDNAMES)
-        writer.writeheader()
-
-    for revision, sqlite, settings in repository.inspect(last_update):
-        record = {
-            **dataclasses.asdict(revision),
-            **dataclasses.asdict(sqlite),
-            **dataclasses.asdict(settings),
-        }
-        record.pop("sqlites")
-        record.pop("settings")
-        if not quiet:
-            writer.writerow({x: record[x] for x in INSPECT_CSV_FIELDNAMES})
-
-    inspection_path.mkdir(exist_ok=True)
-    with (inspection_path / f"{repository.slug}.json").open("w") as f:
-        json.dump(
-            repository,
-            f,
-            indent=indent,
-            default=custom_json_serializer,
-        )
+    application.inspect(
+        ctx.obj["base_path"],
+        ctx.obj["inspection_path"],
+        slug,
+        last_update,
+        click.get_text_stream("stdout") if not quiet else None,
+    )
 
 
 @main.command()
-@click.option(
-    "-i",
-    "--indent",
-    type=int,
-    default=4,
-    help="The indentation in case of JSON output",
+@click.argument(
+    "slug",
+    type=str,
 )
 @click.option(
     "-q/-nq",
@@ -189,53 +143,14 @@ def inspect(ctx, indent, last_update, quiet):
     default=False,
 )
 @click.pass_context
-def plan(ctx, indent, quiet):
+def plan(ctx, slug, quiet):
     """Plans schematisation migration for given inspect result"""
-    repository_slug = ctx.obj["repository"].slug
-    inspection_path = ctx.obj["inspection_path"]
-    if ctx.obj["metadata_path"]:
-        metadata = load_metadata(ctx.obj["metadata_path"])
-    else:
-        metadata = None
-
-    with (inspection_path / f"{repository_slug}.json").open("r") as f:
-        repository = json.load(f, object_hook=custom_json_object_hook)
-
-    assert repository.slug == repository_slug
-
-    result = repository_to_schematisations(repository, metadata)
-    if not quiet:
-        print(f"Schematisation count: {result['count']}")
-
-        for schematisation in result["schematisations"]:
-            revisions = schematisation.revisions
-            rev_rng = f"{revisions[-1].revision_nr}-{revisions[0].revision_nr}"
-            print(f"{schematisation.concat_name}: {rev_rng}")
-
-        print(
-            f"File count: {result['file_count']}, Estimated size: {result['file_size_mb']} MB"
-        )
-
-    with (inspection_path / f"{repository.slug}.plan.json").open("w") as f:
-        json.dump(
-            result,
-            f,
-            indent=indent,
-            default=custom_json_serializer,
-        )
-
-
-@main.command()
-@click.argument(
-    "revision_hash",
-    type=str,
-    required=True,
-)
-@click.pass_context
-def checkout(ctx, revision_hash):
-    """Lists revisions in a repository"""
-    repository = ctx.obj["repository"]
-    repository.checkout(revision_hash)
+    application.plan(
+        ctx.obj["inspection_path"],
+        slug,
+        ctx.obj["metadata_path"],
+        quiet,
+    )
 
 
 @main.command()
@@ -298,11 +213,11 @@ def batch(ctx, remote, uuid, indent, last_update, cache, filters):
         if filters and not re.match(filters, slug):
             continue
         try:
-            download_inspect_plan(
+            application.download_inspect_plan(
                 base_path,
                 inspection_path,
                 metadata,
-                _metadata.slug,
+                slug,
                 remote,
                 uuid,
                 indent,
@@ -313,10 +228,7 @@ def batch(ctx, remote, uuid, indent, last_update, cache, filters):
             logger.warning(f"Could not process {_metadata.slug}: {e}")
         finally:
             # Always cleanup
-            repository = Repository(base_path, slug)
-            if repository.path.exists():
-                shutil.rmtree(repository.path)
-            clear_largefiles_cache()
+            application.delete(base_path, slug)
 
 
 if __name__ == "__main__":